usb: dwc3: Change addr and length declaration to long at dwc3_flush_cache() function
In the dwc3 driver, calls of dwc3_flush_cache() pass addr and length as
longs. Unfortunately, this function converts long on 64 bit machines
to 32 bit int.

This causes "Synchronous Abort" exceptions on 64 bit machines.

To alleviate this problem we accept long arguments to dwc3_flush_cache()
and then explicitly covert them to unsigned types required by
flush_dcache_range().

+++ drivers/usb/dwc3/io.h
@@ -48,8 +48,10 @@ static inline void dwc3_writel(void __iomem *base, u32 offset, u32 value)
 	writel(value, base + offs);
 }
 
-static inline void dwc3_flush_cache(int addr, int length)
+static inline void dwc3_flush_cache(long addr, long length)
 {
-	flush_dcache_range(addr, addr + ROUND(length, CACHELINE_SIZE));
+	flush_dcache_range((unsigned long) addr,
+			   (unsigned long) addr
+			   + ROUND(length, CACHELINE_SIZE));
 }
 #endif /* __DRIVERS_USB_DWC3_IO_H */
